feature: add dynamic settings

+++ particlelife/src/Particle.swift
@@ -11,5 +11,5 @@ import simd
 struct Particle {
     var position: simd_float2
     var velocity: simd_float2
-    var species: Int16
+    var species: Int32
 }
